tests/ProtoR.Domain.UnitTests/SchemaGroupAggregateTests/RulesTests/ProtoBufRulesTests/OneOfAddedRuleTests.cs
namespace ProtoR.Domain.UnitTests.SchemaGroupAggregateTests.RulesTests.ProtoBufRulesTests
{
    using ProtoR.Domain.SchemaGroupAggregate;
    using ProtoR.Domain.SchemaGroupAggregate.Rules.ProtoBufRules;
    using ProtoR.Domain.SchemaGroupAggregate.Schemas;
    using ProtoR.Domain.UnitTests.SchemaFixtures;
    using Xunit;

    public class OneOfAddedRuleTests
    {
        private OneOfAddedRule rule;

        public OneOfAddedRuleTests()
        {
            this.rule = new OneOfAddedRule();
        }

        [Theory]
        [RuleTestData("OneOfAdded", "AddedOneOf")]
        public void Validate_WithAddedOneOf_ShouldNotPass(ProtoBufSchema a, ProtoBufSchema b)
        {
            ValidationResult result = this.rule.Validate(a, b);

            Assert.False(result.Passed);
        }

        [Theory]
        [RuleTestData("OneOfAdded", "InnerOneOf")]
        public void Validate_WithAddedNestedOneOf_ShouldNotPass(ProtoBufSchema a, ProtoBufSchema b)
        {
            ValidationResult result = this.rule.Validate(a, b);

            Assert.False(result.Passed);
        }

        [Theory]
        [RuleTestData("OneOfAdded", "NoAddedOneOf")]
        public void Validate_WithNoAddedOneOf_ShouldPass(ProtoBufSchema a, ProtoBufSchema b)
        {
            ValidationResult result = this.rule.Validate(a, b);

            Assert.True(result.Passed);
        }
    }
}
